refactor(macos): extract InstantMix from AppModel into AppModel+InstantMix.swift
